Restructuring main pape UI
