+++ Content/css/register.css
@@ -1,0 +1,234 @@
+﻿html {
+    scroll-behavior: smooth;
+}
+
+*,
+::after,
+::before {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
+
+body {
+    font-family: 'Rubik', sans-serif;
+    overflow-x: hidden;
+    margin: auto;
+}
+
+ul {
+    list-style-type: none;
+}
+
+a {
+    text-decoration: none;
+}
+
+img:not(.logo) {
+    width: 100%;
+}
+
+img {
+    display: block;
+}
+
+h1,
+h2,
+h3,
+h4,
+h5,
+h6 {
+    font-family: 'Rubik', sans-serif;
+}
+
+a,
+p,
+span,
+button,
+textarea,
+input,
+select,
+option {
+    font-family: 'Mitr', sans-serif;
+}
+
+    input,
+    textarea,
+    button:focus {
+        outline: 0;
+        box-shadow: none;
+    }
+
+section.register {
+    min-height: 100vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+}
+
+    section.register .register-container {
+        background: #d35555;
+        padding: 1rem;
+    }
+
+        section.register .register-container .register-form {
+            background: #fff;
+            padding: 1rem;
+            border-radius: 10px;
+        }
+
+            section.register .register-container .register-form .register-header {
+                display: flex;
+                justify-content: space-between;
+            }
+
+                section.register .register-container .register-form .register-header .back-home a {
+                    color: #000;
+                    font-size: 18px;
+                    font-weight: 500;
+                }
+
+                section.register .register-container .register-form .register-header .report-message {
+                    background: #d35555;
+                    color: #fff;
+                    padding: 0.5rem;
+                    width: 50%;
+                }
+
+                    section.register .register-container .register-form .register-header .report-message p {
+                        font-size: 18px;
+                        font-weight: 600;
+                        color: #fff;
+                    }
+
+                        section.register .register-container .register-form .register-header .report-message p.p2 {
+                            font-size: 15px;
+                            color: #fff;
+                        }
+
+                    section.register .register-container .register-form .register-header .report-message i {
+                        font-size: 18px;
+                        color: #fff;
+                        text-align: center;
+                    }
+
+            section.register .register-container .register-form .register-title h1 {
+                color: #d35555;
+                font-size: 35px;
+                font-weight: 800;
+                line-break: 10px;
+                margin-top: -4rem;
+            }
+
+
+            /* register Form  */
+            section.register .register-container .register-form .form-container {
+                display: flex;
+                justify-content: space-between;
+                align-items: center;
+            }
+
+                section.register .register-container .register-form .form-container form .form-field {
+                    padding-top: 0.8rem;
+                }
+
+                    section.register .register-container .register-form .form-container form .form-field label {
+                        font-size: 18px;
+                        color: #d35555;
+                        padding: 0.5rem 0;
+                    }
+
+                    section.register .register-container .register-form .form-container form .form-field input {
+                        font-size: 18px;
+                        color: #000;
+                        border: none;
+                        border: 1px solid #d35555;
+                        width: 100%;
+                        padding: 5px 10px;
+                    }
+
+                        section.register .register-container .register-form .form-container form .form-field input:focus {
+                            outline: 0;
+                            box-shadow: none;
+                            border: none;
+                            border: 2px solid #d35555;
+                        }
+
+                    section.register .register-container .register-form .form-container form .form-field button.register-button {
+                        width: 100%;
+                        font-weight: 600;
+                        border: none;
+                        background: #d35555;
+                        color: #ffff;
+                        font-size: 20px;
+                        padding: 5px 20px;
+                        border-radius: 5px;
+                        transition: 0.4s;
+                    }
+
+                        section.register .register-container .register-form .form-container form .form-field button.register-button:hover {
+                            background: #da4343 !important;
+                        }
+
+                        section.register .register-container .register-form .form-container form .form-field button.register-button:focus {
+                            outline: 0;
+                            box-shadow: none;
+                            background: #e29090 !important;
+                        }
+
+                section.register .register-container .register-form .form-container .register {
+                    padding-top: 1rem;
+                    text-align: center;
+                }
+
+                    section.register .register-container .register-form .form-container .register p {
+                        font-size: 15px;
+                        color: #000;
+                        margin-bottom: -0.2rem;
+                    }
+
+                    section.register .register-container .register-form .form-container .register a {
+                        font-size: 15px;
+                        color: #da4343;
+                    }
+
+
+                section.register .register-container .register-form .form-container .register-right img {
+                    width: 100%;
+                }
+
+
+
+/* Responsive Section  */
+@media screen and (max-width: 668px) {
+    section.register .register-container .register-form .register-header {
+        flex-direction: column;
+    }
+
+        section.register .register-container .register-form .register-header .report-message {
+            width: 100%;
+        }
+
+    section.register .register-container .register-form .register-title h1 {
+        margin-top: 1rem;
+    }
+}
+
+@media screen and (max-width: 500px) {
+    section.register .register-container .register-form .form-container {
+        flex-direction: column;
+    }
+
+        section.register .register-container .register-form .form-container .register-field {
+            order: calc(2);
+        }
+
+        section.register .register-container .register-form .form-container .register-right {
+            order: calc(1);
+        }
+
+            section.register .register-container .register-form .form-container .register-right img {
+                max-width: 100%;
+            }
+}
